Addresses the most serious part of Ticket #101. Although the form of file reading used ought to self-close the file in question, it does not seem to. Perhaps it is reliant on garbage collection of the temporary File object, but garbage collection occurs so infrequently that the file can stay open for a very long time - so long in fact that "Too many open file" errors are seen during numerous visits to the web site in a short space of time.
In any event, although more verbose, the new code in the Gem is quite
explicit about closing the file after reading and, according to lsof,
does so imediately. As an efficiency note, it is clear that this
aspect of Hub needs a lot of work - the file gets opened and closed
very often (typically twice for a site-based access as the secure
cookie is first read, then written after last accessed time update).